refactor(client,config,tm): move tm config code into tm module
1. revert pkg/config/client_config.go .
2. tm config ut .
